fix params min data in leaf

+++ optimizar/optimizar_lgbm.py
@@ -48,7 +48,7 @@ def objective(trial):
     params_objetivo = {
     'num_leaves' : trial.suggest_int('num_leaves', 20, 10000),
     'learning_rate' : trial.suggest_float('learning_rate', 0.001, 0.4), # mas bajo, más iteraciones necesita
-    'min_data_in_leaf' : p_min_data_in_leaf*n_train_rows,
+    'min_data_in_leaf' : int(p_min_data_in_leaf*n_train_rows),
     'feature_fraction' : trial.suggest_float('feature_fraction', 0.005, .9),
     'feature_fraction_bynode' : trial.suggest_float('feature_fraction_bynode', 0.05, .9), 
     'drop_rate': trial.suggest_float('drop_rate', 0.005, 0.3),
